refactor: extract HeaderDropdown into a dedicated component
