+++ web/src/components/HeaderDropdown.stories.tsx
@@ -1,0 +1,20 @@
+import type { Meta, StoryObj } from '@storybook/react-vite';
+import { HeaderDropdown } from './HeaderDropdown';
+
+const meta: Meta<typeof HeaderDropdown> = {
+  title: 'Components/HeaderDropdown',
+  component: HeaderDropdown,
+  tags: ['autodocs'],
+  decorators: [
+    (Story) => (
+      <div style={{ position: 'relative', height: '400px', width: '300px' }}>
+        <Story />
+      </div>
+    ),
+  ],
+};
+
+export default meta;
+type Story = StoryObj<typeof HeaderDropdown>;
+
+export const Default: Story = {};
