+++ web/src/components/HeaderMenu.scss
@@ -53,118 +53,6 @@
   }
 }
 
-.header-dropdown-card {
-  position: absolute;
-  right: 0;
-  top: 100%;
-  margin-top: 0.5rem;
-  width: 240px;
-  background-color: var(--bg);
-  border: 1px solid var(--border);
-  border-radius: 8px;
-  box-shadow: var(--shadow);
-  padding: 0.5rem 0;
-  z-index: 100;
-  display: flex;
-  flex-direction: column;
-}
-
-.header-menu-section {
-  padding: 0 0.5rem 0.5rem 0.5rem;
-
-  &:not(:last-child) {
-    border-bottom: 1px solid var(--border);
-    margin-bottom: 0.5rem;
-  }
-
-  &:not(:first-child) {
-    padding-top: 0.5rem;
-  }
-}
-
-.header-menu-label {
-  display: block;
-  padding: 0.25rem 0.75rem;
-  text-transform: uppercase;
-  font-size: 0.7rem;
-  font-weight: 600;
-  color: var(--text-muted);
-  letter-spacing: 0.05em;
-}
-
-.header-menu-link {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0.5rem 0.75rem;
-  font-size: 0.875rem;
-  color: var(--text);
-  text-decoration: none;
-  border-radius: 6px;
-  transition:
-    background-color 0.2s ease,
-    color 0.2s ease;
-
-  &:hover {
-    background-color: var(--panel);
-    color: var(--primary);
-  }
-
-  .meta-left {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-  }
-
-  svg {
-    font-size: 1rem;
-  }
-}
-
-.header-meta-link {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0.4rem 0.75rem;
-  font-size: 0.8rem;
-  color: var(--text-muted);
-  text-decoration: none;
-  border-radius: 6px;
-  transition:
-    background-color 0.2s ease,
-    color 0.2s ease;
-
-  &:hover {
-    background-color: var(--panel);
-    color: var(--primary);
-
-    .header-version-tag {
-      background-color: var(--border);
-    }
-  }
-
-  .meta-left {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    flex-shrink: 0;
-  }
-}
-
-.header-version-tag {
-  font-family: var(--mono);
-  background-color: var(--code-bg);
-  color: var(--text-muted);
-  padding: 0.1rem 0.4rem;
-  border-radius: 4px;
-  font-size: 0.7rem;
-}
-
-.header-link-arrow {
-  color: var(--text-muted);
-  font-size: 0.7rem;
-}
-
 @media (prefers-color-scheme: dark) {
   .header-sponsor-btn {
     background-color: rgba(218, 55, 81, 0.1);
